requirements.txt
pygame
RPi.GPIO
uvicorn
fastapi
requests